git-svn-id: svn:///U:/Repos/tep32@335 03528888-2118-ae43-a319-59641b731ea3

+++ branches/reaktivka_001/TepCommon/HTepUsers.cs
@@ -296,7 +296,7 @@ namespace TepCommon
                 {
                     foreach (DataRow r in tblRes.Rows)
                     {
-                        id_alg = (Int32)r[@"ID"];
+                        id_alg = (Int16)r[@"ID"];
 
                         ratio = s_iRatioDefault;
                         round = s_iRoundDefault;
